separate out auth screens styles
