+++ src/screens/Register.js
@@ -1,8 +1,8 @@
 import React, { useState, useLayoutEffect } from 'react';
-import { StyleSheet, SafeAreaView, Text, Image, TextInput, Button, View } from 'react-native';
+import { SafeAreaView, Text, Image, TextInput, Button, View } from 'react-native';
 import { firebase } from '../firebase/config';
 import {resetPassword} from './ResetPassword'
-
+import styles from '../styles/authStyles'
 
 export default function Register({navigation}) {
 	 const [email, setEmail] = useState("");
@@ -52,19 +52,19 @@ export default function Register({navigation}) {
     return (
       <SafeAreaView style={styles.container}>
 				<Image style={styles.image} source={require('../../assets/images/logo.png')} />
-				<View>
+				<View style={{marginBottom: 30}}>
 					<Text>{error}</Text>
 					<View>
 						<Text style={styles.label}>Email Address</Text>
-						<TextInput style={styles.input} placeholder ="Enter email" onChangeText={(email) => setEmail(email)} />
+						<TextInput style={styles.input} autoCorrect={false} autoCapitalize={'none'} onChangeText={(email) => setEmail(email.trim())} />
 					</View>
 					<View>
 						<Text style={styles.label}>Password</Text>
-						<TextInput style={styles.input} secureTextEntry={true} placeholder="Enter password" onChangeText={(password) => setPassword(password)} />
+						<TextInput style={styles.input} secureTextEntry={true} onChangeText={(password) => setPassword(password)} />
 					</View>	
 					<View>
 						<Text style={styles.label}>Confirm Password </Text>
-						<TextInput style={styles.input} secureTextEntry={true} placeholder="Confirm password" onChangeText={(password) => setConfirmPassword(password)} />
+						<TextInput style={styles.input} secureTextEntry={true} onChangeText={(password) => setConfirmPassword(password)} />
 					</View>				
 					<Text style={styles.forgot} onPress={() => {resetPassword(email)}}>Forgot password or email?</Text>	
 					<Button style={styles.button} onPress={() => onSignUp()} title="Sign Up" />
@@ -73,42 +73,3 @@ export default function Register({navigation}) {
     )
 }
 
-const styles = StyleSheet.create({
-	container: {
-    flex: 1,
-		alignItems: "center",
-		justifyContent: "center",
-		padding: 100,
-		// fontFamily: 'Roboto',
-	},
-	image : {
-		width: 300, 
-		height: 300, 
-	},
-	label : {
-		fontSize: 16,
-		color: '#999999',
-		// textShadow: '1px 0px #888888',
-	},
-	input : {
-		// fontFamily: 'Roboto',
-		height: 40,
-		fontSize: 16,
-		// marginTop: 8,
-		marginVertical: 10,
-		paddingLeft: 3,
-		paddingRight: 3,
-		color: '#000000',
-		borderColor: '#9b9b9b',
-		borderBottomWidth: 2,	
-	},
-	forgot : {
-		marginBottom: 15,
-		width: 300,
-		color: '#147DF0'		
-	},
-	button: {
-		width: 300,
-		backgroundColor: '#147DF0',
-	}
-});
